Easy Qvarn setup for local development (WIP)
Started so work to package Qvarn in a way, that it would be really easy to run
for development.

The idea is to build a python package with a console script for running uwsgi
and haproxy under honcho management. Then in order to run Qvarn one should do:

    $ sudo apt install postgresql-server haproxy
    $ pip install qvarntesting
    $ qvtestserver 0.81

Currently I did some structure refactoring, added setup.py and started to work
on qvtestserver script.

TODO:

- Finish qvtestserver configuration templates.

- Automatically create postgresql databose if missing.

- Finish qvtestserver process management using honcho.

- Maybe possibility to configure database, for example, option to Qvarn on
  SQLite.